+++ src/helm/benchmark/heim_run_specs.py
@@ -1,12 +1,12 @@
 from typing import List, Optional
 
 from helm.common.image_generation_parameters import ImageGenerationParameters
-from .adaptation.adapter_spec import AdapterSpec
-from .adaptation.adapters.adapter_factory import ADAPT_GENERATION
-from .metrics.metric import MetricSpec
-from .run_specs import run_spec_function, get_basic_metric_specs
-from .runner import RunSpec
-from .scenarios.scenario import ScenarioSpec
+from helm.benchmark.adaptation.adapter_spec import AdapterSpec
+from helm.benchmark.adaptation.adapters.adapter_factory import ADAPT_GENERATION
+from helm.benchmark.metrics.metric import MetricSpec
+from helm.benchmark.run_specs.classic_run_specs import run_spec_function, get_basic_metric_specs
+from helm.benchmark.runner import RunSpec
+from helm.benchmark.scenarios.scenario import ScenarioSpec
 
 
 ############################################################
