+++ src/helm/benchmark/scenarios/test_scenario.py
@@ -1,4 +1,4 @@
-from helm.benchmark.run_specs import get_scenario_spec_tiny
+from helm.benchmark.run_specs.classic_run_specs import get_scenario_spec_tiny
 from helm.benchmark.scenarios.scenario import create_scenario, Scenario, Input, PassageQuestionInput
